seongeun, refactor: Contract Validator 패키지 분리 #133
