+++ src/main/java/com/househub/backend/domain/contract/dto/CreateContractResDto.java
@@ -1,5 +1,6 @@
 package com.househub.backend.domain.contract.dto;
 
+import com.househub.backend.domain.contract.entity.Contract;
 import lombok.Getter;
 
 @Getter
@@ -9,7 +10,7 @@ public class CreateContractResDto {
         this.id = id;
     }
 
-    public static CreateContractResDto toDto(Long id) {
-        return new CreateContractResDto(id);
+    public static CreateContractResDto fromEntity(Contract contract) {
+        return new CreateContractResDto(contract.getId());
     }
 }
