+++ src/main/java/com/househub/backend/domain/contract/service/ContractReader.java
@@ -15,8 +15,6 @@ public interface ContractReader {
 
 	Page<Contract> findPageBySearchDto(ContractSearchDto searchDto, Pageable pageable, Long agentId); // 계약 검색
 
-	void validateNoInProgressContract(Customer customer, Property property); // 진행중인 계약이 있는지 확인
-
 	Page<Contract> findContractsByAgentAndCustomer(Long agentId, Long customerId, Pageable pageable);
 
 	Page<Contract> findContractsByProperties(Long agentId, List<Property> propertyIds, Pageable pageable);
