+++ src/main/java/com/househub/backend/domain/contract/service/impl/ContractServiceImpl.java
@@ -1,5 +1,6 @@
 package com.househub.backend.domain.contract.service.impl;
 
+import com.househub.backend.domain.contract.validator.ContractValidator;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -39,6 +40,7 @@ public class ContractServiceImpl implements ContractService {
 	private final ContractStore contractStore;
 	private final PropertyReader propertyReader;
 	private final CustomerReader customerReader;
+	private final ContractValidator contractValidator;
 
 	/**
 	 * 계약 등록
@@ -52,28 +54,24 @@ public class ContractServiceImpl implements ContractService {
 		// 계약할 매물 조회
 		Property property = propertyReader.findByIdOrThrow(dto.getPropertyId(), agentDto.getId());
 		Customer customer = null;
-
 		if (dto.getCustomerId() != null) {
 			// 계약자 존재할 경우 고객 조회
 			customer = customerReader.findByIdOrThrow(dto.getCustomerId(), agentDto.getId());
 			// 매물을 등록한 고객과 계약할 고객이 동일한 경우 예외 처리
 			validateCustomerIsNotPropertyOwner(customer, property);
 			// 같은 계약자가 동일한 매물에 대해 진행중인 계약이 있는지 확인
-			contractReader.validateNoInProgressContract(customer, property);
+			contractValidator.validateNoInProgressContract(customer, property);
 		}
 		// dto → entity 변환 후 저장
 		Agent agent = agentDto.toEntity();
 		Contract contract = dto.toEntity(property, customer, agent);
-		// 계약 가능 상태로 등록할 경우, 매물과 계약 상태 enable
+		// 계약 가능 상태로 등록할 경우, 매물 활성화 상태 enable
 		if(dto.getContractStatus() == ContractStatus.AVAILABLE) {
 			property.enable();
-			contract.enable();
-		} else {
-			contract.disable();
 		}
 		contract = contractStore.create(contract);
 		// 응답 객체 리턴
-		return CreateContractResDto.toDto(contract.getId());
+		return CreateContractResDto.fromEntity(contract);
 	}
 
 	/**
@@ -91,14 +89,15 @@ public class ContractServiceImpl implements ContractService {
 		}
 		// 매물 조회
 		Property property = contract.getProperty();
-		// 고객을 설정한 경우, 검증
+		// 계약자를 수정한 경우, 검증
 		if(dto.getCustomerId() != null) {
 			// 고객 조회
 			Customer customer = customerReader.findByIdOrThrow(dto.getCustomerId(), agentDto.getId());
 			// 매물을 등록한 고객과 계약할 고객이 동일한 경우 예외 처리
 			validateCustomerIsNotPropertyOwner(customer, property);
 			// 같은 계약자가 동일한 매물에 대해 진행중인 계약이 있는지 확인
-			contractReader.validateNoInProgressContract(customer, property);
+			contractValidator.validateNoInProgressContract(customer, property);
+			// 계약자 수정
 			contractStore.updateCustomer(contract, customer);
 		}
 		// 계약 정보 수정
@@ -116,15 +115,10 @@ public class ContractServiceImpl implements ContractService {
 	@Transactional(readOnly = true)
 	public ContractListResDto findContracts(ContractSearchDto searchDto, Pageable pageable, AgentResDto agentDto) {
 		Agent agent = agentDto.toEntity();
-		log.info("findContracts searchDto: {}", searchDto.getAgentName());
-		log.info("findContracts agent id: {}", agentDto.getId());
-
 		// 해당 공인중개사가 체결한 계약만 조회
 		Page<Contract> contractPage = contractReader.findPageBySearchDto(searchDto, pageable, agent.getId());
-		log.info("findContracts contractPage: {}", contractPage.getContent().size());
 		// 계약 엔티티를 dto 로 변환하여 리스트로 반환
 		Page<FindContractResDto> response = contractPage.map(FindContractResDto::toDto);
-		log.info("findContracts response size: {}", response.getContent().size());
 		return ContractListResDto.fromPage(response);
 	}
